requirements.txt
blinker==1.6.2
click==8.1.7
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
